requirements.txt
comtypes==1.1.10
keyboard==0.13.5
numpy==1.21.5
opencv-contrib-python==4.5.4.60
Pillow==8.4.0
pytesseract==0.3.8
pywin32==303
pywinauto==0.6.8
six==1.16.0
tld==0.12.6
